feat: replace terminal activities with matching page
- Replace "Toiminnot" navigation with "Sovittaminen" (Matching)
- Create new two-column matching layout showing donations and requests side-by-side
- Move comprehensive activities functionality to dashboard page with analytics cards and filtering
- Remove route displays that aren't developed yet
- Fix rounded corner styling for matching page headers
- Update Finnish translations for matching terminology
